split adminHome progress charts into their own component
